feat: perform rag indexing prior to run rather than as-we-run
This helps with display, so that we can have a single indicatif MultiProgress

Eventually, we may want to move this, as well as the retrieval, to `plan()`. This would give us a nice split between rag logic and execution logic. The `run()` would then be completely separate from any RAG logic. Eventually this might help us separate the planner and the executor even into two separate services, etc.

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>