fix: span tokenization fixes to properly pad and support cropping Assistant msgs
Previously we were doing the span tokenization in two passes, inserting padding in both steps (??). With this change, we collect the tokens as we scan the span query tree, which allows us to know "where we are" in the token stream, and thus allows us to insert padding (or cropping) as we go.

Cropping of assistant messages is currently needed because VLLM scans for cache matches from the front of the token sequence. As soon as it finds a cache miss, it gives up on finding any more cache hits. Assistant *output* is not padding (because this is generated by vllm itself), whereas assistant *input* ... if padded may thus lead to a cache miss in the middle of the stream. This can occur if passing multiple assistant outputs in a subsequent input.

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>